test: move tests to adapters/REST
