+++ test/unit/entities/team-membership-test.js
@@ -12,7 +12,6 @@ import {
   failingActionTest,
   entityCollectionWrappedTest,
 } from '../test-creators/instance-entity-methods'
-import { expect } from 'chai'
 
 function setup(promise) {
   return {
@@ -40,33 +39,6 @@ describe('Entity TeamMembership', () => {
     })
   })
 
-  /**
-   * Move to adapters/REST
-   */
-  test.skip('TeamMembership update', async () => {
-    const { httpMock, entityMock } = setup()
-    entityMock.sys.version = 2
-    entityMock.sys.team = { sys: { id: 'team1' } }
-    const entity = wrapTeamMembership(httpMock, entityMock)
-    entity.admin = true
-    return entity.update().then((response) => {
-      expect(response.toPlainObject, 'response is wrapped').to.be.ok
-      expect(httpMock.put.args[0][0]).equals(
-        `/organizations/org-id/teams/team1/team_memberships/${entityMock.sys.id}`,
-        'url is correct'
-      )
-      expect(httpMock.put.args[0][2].headers['X-Contentful-Version']).equals(
-        2,
-        'version header is sent'
-      )
-      return {
-        httpMock,
-        entityMock,
-        response,
-      }
-    })
-  })
-
   test('TeamMembership update fails', async () => {
     return failingActionTest(setup, {
       wrapperMethod: wrapTeamMembership,
@@ -77,27 +49,6 @@ describe('Entity TeamMembership', () => {
   test('TeamMembership delete', async () => {
     return entityDeleteTest(setup, {
       wrapperMethod: wrapTeamMembership,
-    })
-  })
-
-  /**
-   * Move to adapters/REST
-   */
-  test.skip('TeamMembership delete', async () => {
-    const { httpMock, entityMock } = setup()
-    entityMock.sys.version = 2
-    entityMock.sys.team = { sys: { id: 'team1' } }
-    const entity = wrapTeamMembership(httpMock, entityMock)
-    return entity.delete().then((response) => {
-      expect(httpMock.delete.args[0][0]).equals(
-        `/organizations/org-id/teams/team1/team_memberships/${entityMock.sys.id}`,
-        'url is correct'
-      )
-      return {
-        httpMock,
-        entityMock,
-        response,
-      }
     })
   })
