+++ test/unit/entities/team-test.js
@@ -9,7 +9,6 @@ import {
   entityUpdateTest,
   entityDeleteTest,
 } from '../test-creators/instance-entity-methods'
-import { expect } from 'chai'
 
 function setup(promise) {
   return {
@@ -37,32 +36,6 @@ describe('Entity TeamSpaceMembership', () => {
     })
   })
 
-  /**
-   * Move to adapters/REST
-   */
-  test.skip('Team update', async () => {
-    const { httpMock, entityMock } = setup()
-    entityMock.sys.version = 2
-    const entity = wrapTeam(httpMock, entityMock)
-    entity.description = 'new description'
-    return entity.update().then((response) => {
-      expect(response.toPlainObject, 'response is wrapped').to.be.ok
-      expect(httpMock.put.args[0][0]).equals(
-        `/organizations/org-id/teams/${entityMock.sys.id}`,
-        'url is correct'
-      )
-      expect(httpMock.put.args[0][2].headers['X-Contentful-Version']).equals(
-        2,
-        'version header is sent'
-      )
-      return {
-        httpMock,
-        entityMock,
-        response,
-      }
-    })
-  })
-
   test('Team update fails', async () => {
     return failingActionTest(setup, {
       wrapperMethod: wrapTeam,
@@ -73,26 +46,6 @@ describe('Entity TeamSpaceMembership', () => {
   test('Team delete', async () => {
     return entityDeleteTest(setup, {
       wrapperMethod: wrapTeam,
-    })
-  })
-
-  /**
-   * Move to adapters/REST
-   */
-  test.skip('Team delete', async () => {
-    const { httpMock, entityMock } = setup()
-    entityMock.sys.version = 2
-    const entity = wrapTeam(httpMock, entityMock)
-    return entity.delete().then((response) => {
-      expect(httpMock.delete.args[0][0]).equals(
-        `/organizations/org-id/teams/${entityMock.sys.id}`,
-        'url is correct'
-      )
-      return {
-        httpMock,
-        entityMock,
-        response,
-      }
     })
   })
